Increase block_num counter size to u64 since it may overflow on very big files. Temporarily enable overflow check for testing.

+++ gbam_tools/src/compressor.rs
@@ -14,7 +14,7 @@ use lzzzz::lz4;
 use crate::writer::BlockInfo;
 
 pub(crate) enum OrderingKey {
-    Key(u32),
+    Key(u64),
     UnusedBlock,
 }
